# swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'Cliente API'
  description: 'Esta API tem o objetivo cadastrar, consultar, alterar e excluir clientes
  [Saiba mais](https://www.google.com.br)
   '
consumes:
  - application/json
produces:
  - application/json
 
tags:
  - name: 'Consulta'
  - name: 'Cadastro'
  - name: 'Manutencao'
  - name: 'Exclusao'
  - name: 'API'
 
host: localhost:8080
basePath: /cliente-api/v1
schemes:
  - http
  - https
paths:
#localhost:8080//cliente-api/v1/clientes (consulta todos os clientes)
  /clientes:
    get:
      tags:
        - Consulta
      summary: 'Consulta todos clientes'
      operationId: 'consultaTodos'
      description: 'Essa operação tem como objetivo consultar todos os clientes'
      responses:
        200:
          description: 'Sucesso ao consultar todos clientes.'
          schema:
            $ref: '#/definitions/Clientes'
        400:
          description: 'Consulta invalida'
        401:
          description: 'Consulta nao autorizada'
        404:
          description: 'Nenhum cliente encontrado para a consulta'
        500:
          description: 'Erro interno ao tentar realizar a consulta'
         
         
  # localhost:8080//cliente-api/v1/cliente
  /cliente:
    post:
      tags:
        - Cadastro
      summary: "Cadastro de um novo cliente."
      operationId: "cadastraNovo"
      description: "Esta operacao tem por objetivo cadastrar um novo cliente"
      parameters:
        - in: "body"
          name: "cliente"
          required: true
          schema:
           $ref: "#/definitions/Cliente"
      responses:
        201:
          description: "Sucesso ao cadastrar um novo cliente"
          schema:
            $ref: "#/definitions/Cliente"
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado."
         
        400:
          description: "Cadastro invalido"
        401:
          description: "Cadastro nao autorizado"
        500:
          description: "Erro ao realizar o cadastro"
 
  # localhost:8080//cliente-api/v1/cliente/123
  /cliente/{id}:
    put:
      tags:
        - Manutencao
      summary: "Altera um cliente existente"
      operationId: "alteraExistente"
      description: "Esta operação tem por objetivo alterar um cliente."
      parameters:
      - in: "path"
        name: "id"
        type: "integer"
        format: "int32"
        required: true
        description: "Id do cliente."
      - in: "body"
        name: "cliente"
        required: true
        schema:
          $ref: "#/definitions/Cliente"
      responses:
        202:
          description: "Sucesso ao alterar um cliente existente."
          schema:
            $ref: "#/definitions/Cliente"
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado."
        400:
          description: "Alteração inválida."
        401:
          description: "Alteração não autorizada."
        500:
          description: "Erro interno ao tentar realizar a alteração."
    
    get:
      tags:
      - Consulta
      summary: "Consulta um cliente existe por id."
      operationId: "consultaPorId"
      description: "Esta consulta tem por objetivo consultar um cliente existente por id"
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          format: "int32"
          required: true
          description: "Numero do id do Cliente"
      responses:
        200:
          description: "Sucesso ao consultar cliente por id."
          schema:
            $ref: "#/definitions/Cliente"
        400:
          description: "Consulta invalida" 
        401:
          description: "Consulta nao autorizada"
        404:
          description: "Nenhum cliente encontrado para a consulta por id."
        500:
          description: "Erro interno."
          
    delete:
      tags:
        - Exclusao
      summary: "Exclui um cliente existente."
      operationId: "ExcluiExistente"
      description: "Esta operação tem por objetivo excluir um cliente existente."
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          required: true
          format: "int32"
      responses:
          204:
            description: "Sucesso ao excluir um cliente existente."
          400:
            description: "Exclusão inválida."
          401:
            description: "Exclusão não autorizada."
          500:
            description: "Erro interno ao tentar realizar a exclusão."
            
  /cliente/consulta/{sobrenome}:
    get:
      tags:
        - Consulta
      summary: "Consulta cliente existente por sobrenome."
      operationId: "consultaPorSobrenome"
      description: "Esta operação tem por objetivo consultar um cliente existente por sobrenome."
      parameters:
        - in: "path"
          name: "sobrenome"
          type: "string"
          required: true
          description: "Sobrenome do Cliente."
      responses:
        200:
          description: "Sucesso ao consultar cliente por sobrenome."
          schema:
            $ref: "#/definitions/Clientes"
        400:
           description: "Consulta inválida."
        401:
           description: "Consulta não autorizada."
        404:
           description: "Nenhum cliente encontrado para pesquiza."
        500:
           description: "Erro interno ao tentar realizar a consulta por sobrenome."
  
  /cliente/{status}/{id}:
    put:
      tags:
        - Manutencao
      summary: "Alteração de status de cliente por id."
      operationId: "alteraStatusPorId"
      description: "Esta operacao tem por objetivo alterar o status do cliente por id."
      parameters:
        - in: "path"
          name: "status"
          required: true
          type: "string"
          enum: ["ativo", "inativo"]
          description: "Status do cliente."
        - in: "path"
          name: "id"
          type: "integer"
          format: "int32"
          required: true
          description: "Numero do id do cliente."
      responses:
          202:
            description: "Sucesso ao alterar o status de um cliente existente."
            schema:
              $ref: "#/definitions/Cliente"
            headers:
              Location:
                type: "string"
                description: "URI do cliente alterado."
          400:
            description: "Alteração de status invalida."
          401:
           description: "Alteração de status não autorizada."
          500:
            description: "Erro interno ao tentar realizar alteração do status."
  /readme:
    get:
      tags: 
        - API
      summary: "Recupera readme da API"
      operationId: "readme"
      description: "Esta operação retorna o read da API."
      responses:
        200:
          description: "Sucesso ao recuperar o readme da API."
          schema:
            type: "string"
        500:
          description: "Erro interno ao tentar recuperar o readme da API."  
      
definitions:
 
  Clientes:
    type: 'array'
    items:
      $ref: '#/definitions/Cliente'
 
 
  Cliente:
    type: 'object'
    properties:
      id:
        type: 'integer'
        format: 'int32'
        description: 'Identificador do cliente.'
      nome:
        type: 'string'
        description: 'Nome do cliente.'
      sobrenome:
        type: 'string'
        description: 'Sobrenome do cliente.'
      iniciais:
        type: 'string'
        minLength: 2
        maxLength: 7
        description: 'Iniciais do nome completo do cliente.'
      dataNascimento:
        type: 'string'
        format: 'date'
        description: 'Data de nascimento do cliente.'
      dataAtualizacao:
        type: 'string'
        format: 'date-time'
        description: 'Data de atualizacao dos dados do cliente.'
      status:
        type: boolean
        description: 'Status ativo ou inativo.'
      tipo:
        type: 'string'
        enum: ['interno', 'externo']
        
    
      